% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BB.250}
\alias{BB.250}
\title{BB.250 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
BB.250
}
\description{
\itemize{
 \item Target Soil Properties: SOC_target, pH_target, Clay_target
 \item Groups of Features: ER, Gamma, pH_ISE, RSS, DEM
 \item Sample size: 250
 \item Number of Features: 16
 \item Coordinates: With coordinates (EPSG: 25833)
 \item Location: Brandenburg, Germany
 \item Sampling Design: Triangular grid sampling
 \item Study Area Size: 52 ha
 \item Contact Information: SVogel@atb-potsdam.de
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item \strong{Changelog:}
   \itemize{
     \item Version 1.1 (15.08.2024): Updated sampling date for \code{ER} group.
     \item Version 1.0 (02.08.2024): Initial release.
   }
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOC_target}}{Unit: \% \cr
   Protocol: Measuring CO2 release during dry combustion after removing inorganic carbon (DIN ISO 10694)\cr
   Sampling Date: May 2020\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{clay_target}}{Unit: \% \cr
   Protocol: Wet sieving and sedimentation after removal of organic matter (DIN ISO 11277)\cr
   Sampling Date: May 2020\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{pH_target}}{Unit: Unitless \cr
   Protocol: In CaCl\eqn{_2} solution with a glass electrode (DIN ISO 10390)\cr
   Sampling Date: May 2020\cr
   Sampling Depth: 0 - 30 cm} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{ER}}{
  No. Features: 1 \cr
  Feature Code(s): ER \cr
  Unit: \eqn{\Omega} m \cr
  Sensing: Array of multiple rolling electrodes(Geophilus company, Caputh, Germany) on Geophilus platform with 0 - 25 cm measuring depth (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: August 2021
  }

  \item{\code{pH_ISE}}{
  No. Features: 1 \cr
  Feature Code(s): pH_ISE \cr
  Unit: unitless \cr
  Sensing: Soil pH Manager (VERIS Technologies, Salinas, USA ) from VerisMSP3 (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: August 2021
  }

  \item{\code{Gamma}}{
  No. Features: 1 \cr
  Feature Code(s): G_Total_Counts \cr
  Unit: unitless \cr
  Sensing: Gamma sensor on Geophilus platform (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: August 2021
  }

  \item{\code{RSS}}{
  No. Features: 10 \cr
  Feature Code(s): B02, B03, B04, B05, B06, B07, B08, B08A, B11, B12 \cr
  Unit: unitless \cr
  Sensing: Sentinel-2 bare soil Image (Level-2A) from “Copernicus Open Access Hub”, with bands of 10 - 20 m spatial resolution\cr
  Processing: Extracting RSS values from raster at soil sampling locations \cr
  Sensing Date: April 2020
  }
  \item{\code{DEM}}{
  No. Features: 3 \cr
  Feature Code(s): Altitude, Slope, TPI \cr
  Unit: Altitude (m), Slope (°), TPI (unitless) \cr
  Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from “GeoBasis-DE/LGB”\cr
  Processing: Slope and TPI through \code{\link[raster]{terrain}} function of the \pkg{raster} package \cr
  Sensing Date: LiDAR March 2009, images for photogrammetry May 2018
  }
}


\strong{Folds:}
Vector with 250 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 25833)
}
\examples{
# Load the dataset & folds
BB.250

# Access the dataset
BB.250$Dataset

# Access the folds
BB.250$Folds

# Access the coordinates
BB.250$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_BB.250 <- BB.250$Dataset[BB.250$Folds != 1,]
testing_data_BB.250 <- BB.250$Dataset[BB.250$Folds == 1,]

}
\keyword{datasets}
