% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BB.30_2}
\alias{BB.30_2}
\title{BB.30_2 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
BB.30_2
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa, Gamma, RSS, VI
 \item Sample Size: 30
 \item Number of Features: 13
 \item Coordinates: With coordinates (EPSG: 25833)
 \item Location: Brandenburg, Germany
 \item Sampling Design: Regular grid sampling
 \item Study Area Size: 1.4 ha
 \item Geological Setting: Pleistocene young morainic landscape of the Weichselian glaciation with predominantly glacial sand
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Pablo Rosso (Pablo.Rosso@zalf.de), Leibniz Centre for Agricultural Landscape Research (ZALF)
     \item Sebastian Vogel (SVogel@atb-potsdam.de), Leibniz Institute for Agricultural Engineering and Bioeconomy (ATB)
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured CO\eqn{_2} release during dry combustion after removing inorganic carbon with an acid (DIN ISO 10694)
   \item Sampling Date: September 2022
   \item Sampling Depth: 0 - 30 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in CaCl\eqn{_2} suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: September 2022
   \item Sampling Depth: 0 - 30 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water, German adaptation (DIN ISO 11277)
   \item Sampling Date: September 2022
   \item Sampling Depth: 0 - 30 cm
   }
}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from “GeoBasis-DE/LGB”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sampling Date: LiDAR March 2011, images for photogrammetry May 2022
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: Array of multiple rolling electrodes (Geophilus company, Caputh, Germany) on RapidMapper platform with exploration depth of 0 - 50 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date:  September 2022
  }

  Gamma
  \itemize{
  \item Number of Features: 5
  \item Code(s): \code{G_Total_Counts, G_K, G_U, G_Th, G_Cs}
  \item Unit: Unitless
  \item Sensing: Passive gamma sensor (MS-2000-CsI-MTS, Medusa Radiometrics BV, Groningen, Netherlands) on RapidMapper platform, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: September 2022
  }

  RSS – Remote Sensing Derived Spectral Data
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{B04}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 bare soil image (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Extracting RSS values from raster at soil sampling locations, selecting single band due to low sample size
  \item Sampling Date: September 2022
  }
  VI - Vegetation Indices
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{NDVI}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 Image during vegetative period (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{NDVI} as (B08 - B04) / (B08 + B04), extracting VI values from raster at soil sampling locations
  \item Sampling Date: April 2023
   }
}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 25833)
}
\examples{
# Load the dataset list
BB.30_2

# Access the dataset
BB.30_2$Dataset

# Access the folds
BB.30_2$Folds

# Access the coordinates
BB.30_2$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_BB.30_2 <- BB.30_2$Dataset[BB.30_2$Folds != 1,]
testing_data_BB.30_2 <- BB.30_2$Dataset[BB.30_2$Folds == 1,]

}
\keyword{datasets}
