% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BB.72}
\alias{BB.72}
\title{BB.72 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
BB.72
}
\description{
\itemize{
 \item Target Soil Properties: SOC_target, pH_target, Clay_target
 \item Groups of Features: ER, Gamma, ISE.pH, RSS, DEM
 \item Sample size: 72
 \item Number of Features: 16
 \item Coordinates: With coordinates (EPSG: 25833)
 \item Location: Brandenburg, Germany
 \item Sampling Design: Regular Grid Sampling
 \item Study Area: 3.4 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: SVogel@atb-potsdam.de
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOC_target}}{Unit: \% \cr
   Protocol: Dry combustion after removing inorganic carbon (DIN ISO 10694)\cr
   Sampling Date: August 2022\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{clay_target}}{Unit: \% \cr
   Protocol: Wet sieving and sedimentation after removal of organic matter (DIN ISO 11277)\cr
   Sampling Date: August 2022\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{pH_target}}{Unit: Unitless \cr
   Protocol: In CaCl\eqn{_2} solution with a glass electrode (DIN ISO 10390)\cr
   Sampling Date: August 2022\cr
   Sampling Depth: 0 - 30 cm} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{ER}}{
  No. Features: 1 \cr
  Feature Code(s): ER \cr
  Unit: \eqn{\Omega} m \cr
  Sensing: Array of multiple rolling electrodes (Geophilus company, Caputh, Germany) on Geophilus platform with 0 - 25 cm measuring depth (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: July 2020
  }

  \item{\code{pH_ISE}}{
  No. Features: 1 \cr
  Feature Code(s): pH_ISE \cr
  Unit: unitless \cr
  Sensing: Soil pH Manager from VerisMSP3 (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: August 2021
  }

  \item{\code{Gamma}}{
  No. Features: 1 \cr
  Feature Code(s): G_Total_Counts \cr
  Unit: unitless \cr
  Sensing: Gamma sensor on Geophilus platform (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: July 2020
  }

  \item{\code{RSS}}{
  No. Features: 10 \cr
  Feature Code(s): B02, B03, B04, B05, B06, B07, B08, B08A, B11, B12 \cr
  Unit: unitless \cr
  Sensing: Sentinel-2 bare soil Image (Level-2A), with bands of 10 - 20 m spatial resolution\cr
  Processing: Extracting RSS values from raster at soil sampling locations \cr
  Sensing Date: August 2022
  }
  \item{\code{DEM}}{
  No. Features: 3 \cr
  Feature Code(s): Altitude, Slope, TDI \cr
  Unit: Altitude (m), Slope (°), TDI (unitless) \cr
  Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from the “State office for Land Surveying and Geoinformation Brandenburg”\cr
  Processing: Slope and TPI through \code{\link[raster]{terrain}} function of the \pkg{raster} package \cr
  Sensing Date: LiDAR March 2011, Images for photogrammetry May 2022
  }
}


\strong{Folds:}
Vector with 72 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe With coordinates (EPSG: 25833)
}
\examples{
# Load the dataset & folds
BB.72

# Access the dataset
BB.72$Dataset

# Access the folds
BB.72$Folds

# Access the coordinates
BB.72$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_BB.72 <- BB.72$Dataset[BB.72$Folds != 1,]
testing_data_BB.72 <- BB.72$Dataset[BB.72$Folds == 1,]

}
\keyword{datasets}
