% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC.58}
\alias{BC.58}
\title{BC.58 Dataset}
\format{
An object of class \code{list} of length 3.
}
\usage{
BC.58
}
\description{
\itemize{
 \item dataset will probably be dropped because something seems wrong! Performance with MIR data should be much better !
 }
}
\keyword{datasets}
