% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CV.98}
\alias{CV.98}
\title{CV.98 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
CV.98
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: vis-NIR
 \item Sample Size: 98
 \item Number of Features: 2,151
 \item Coordinates: Without coordinates because dataset was not georeferenced
 \item Location: Canton of Vaud, Switzerland
 \item Sampling Design: Stratified random sampling based on different treatments of organic amendments, fertilization, crop rotation and soil cultivation
 \item Study Area Size: 28 ha
 \item Geological Setting: Glacial or fluvioglacial deposits
 \item Previous Data Publication: Full dataset published in Metzger et al. (2024) but under embargo until June 2025
 \item Contact Information:
   \itemize{
     \item Konrad Metzger (konrad.metzger@agroscope.admin.ch), Agroscope
     \item Luca Bragazza (luca.bragazza@agroscope.admin.ch), Agroscope
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured through titration after oxidization of the organic carbon (Walkley & Black 1934)
   \item Sampling Date: May - June and September - October 2021
   \item Sampling Depth: 0 - 20 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (NF ISO 10390)
   \item Sampling Date: May - June and September - October 2021
   \item Sampling Depth: 0 - 20 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: May - June and September - October 2021
   \item Sampling Depth: 0 - 20 cm
   }

}
\cr
Groups of Features:
\describe{
  vis-NIR – Visible and Near Infrared Spectroscopy
  \itemize{
  \item Number of Features: 2,151
  \item Code(s): \code{wl_350, wl_351, wl_352} ... \code{wl_2500}
  \item Unit: \% (Reflectance)
  \item Sensing: vis-NIR spectrometer ( PSR+3500, Spectral Evolution, Haverhill, USA), in-situ from Edelman auger (denoted as scantype = a in original dataset publication), spectral range was 350 - 2500 nm at 3 - 8 nm intervals
  \item Processing: Resampling to 1 nm intervals
  \item Sampling Date: May - June and September - October 2021
  \item Spectral Information (After Data Processing):
   \itemize{
     \item Data Representation: Wavelength (in nm)
     \item Spectral Resolution: 1 nm
     \item Spectral Range: 350 – 2500 nm
   }
  }

}
\cr

\strong{Folds:}
 Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset list
CV.98

# Access the dataset
CV.98$Dataset

# Access the folds
CV.98$Folds

# Access the coordinates but empty for CV.98 (i.e. NA)
CV.98$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_CV.98 <- CV.98$Dataset[CV.98$Folds != 1,]
testing_data_CV.98 <- CV.98$Dataset[CV.98$Folds == 1,]

}
\references{
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.\cr
\cr
Metzger, K., Liebisch, F., Herrera, J. M., Guillaume, T., Walder, F. & Bragazza, L. (2024). Agroscope_SoilSpectralLibrary _2024. Zenodo repository. https://doi.org/10.5281/zenodo.11204174 \cr
\cr
Walkley, A. & Black, I. A. (1934). An examination of the Degtjareff method for determining soil organic matter, and a proposed modification of the chromic acid titration method. Soil science, 37(1), 29-38.
}
\keyword{datasets}
