% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{G.104}
\alias{G.104}
\title{G.104 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
G.104
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, RSS, VI
 \item Sample Size: 104
 \item Number of Features: 16
 \item Coordinates: With coordinates (EPSG: 32722)
 \item Location: Goias, Brazil
 \item Sampling Design: Regular grid sampling
 \item Study Area Size: 95 ha
 \item Geological Setting: Heavily weathered soils originating from sedimentary rocks (claystone and sandstone)
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Domingos Sarvio Magalhaes Valente (valente@ufv.br), Federal University of Vicosa
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): 28.02.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (28.02.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured through titration after oxidization of the organic carbon (Walkley & Black 1934)
   \item Sampling Date: July 2022
   \item Sampling Depth: 0 – 20 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode with a 2.5:1 liquid:soil volumetric ratio
   \item Sampling Date: July 2022
   \item Sampling Depth: 0 – 20 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: July 2022
   \item Sampling Depth: 0 – 20 cm
   }
}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (30 m) based on synthetic aperture radar from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sampling Date: December 2010
  }

  RSS – Remote Sensing Derived Spectral Data
  \itemize{
  \item Number of Features: 12
  \item Code(s): \code{B01, B02, B03, B04, B05, B06, B07, B08, B8A, B09, B11, B12}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 bare soil image (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Extracting RSS values from raster at soil sampling locations
  \item Sampling Date: July 2022
  }

  VI - Vegetation Indices
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{NDVI, GNDVI}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 image during vegetative period (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{NDVI} as (B08 - B04) / (B08 + B04) and \code{GNDVI} as (B08 - B03) / (B08 + B03), extracting VI values from raster at soil sampling locations
  \item Sampling Date: December 2023
  }

}
\cr

\strong{Folds:}
 Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32722)
}
\examples{
# Load the dataset list
G.104

# Access the dataset
G.104$Dataset

# Access the folds
G.104$Folds

# Access the coordinates
G.104$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_G.104 <- G.104$Dataset[G.104$Folds != 1,]
testing_data_G.104 <- G.104$Dataset[G.104$Folds == 1,]

}
\references{
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.\cr
\cr
Walkley, A. & Black, I. A. (1934). An examination of the Degtjareff method for determining soil organic matter, and a proposed modification of the chromic acid titration method. Soil science, 37(1), 29-38.
}
\keyword{datasets}
