% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MG.44}
\alias{MG.44}
\title{MG.44 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
MG.44
}
\description{
\itemize{
 \item Target Soil Properties: SOM_target, pH_target, Cay_target
 \item Groups of Features: vis-NIR
 \item Sample size: 44
 \item Number of Features: 351
 \item Coordinates: With coordinates (EPSG: 32721)
 \item Location: Mato Grosso, Brazil
 \item Sampling Design: Random subsample from a larger grid sampling campaign
 \item Study Area: 13 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: tiagosrt@usp.br
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOM_target}}{XXX}
  \item{\code{clay_target}}{XXX}
  \item{\code{pH_target}}{XXX} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{visNIR}}{XXX
  }
}


\strong{Folds:}
Vector with 44 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32721)
}
\examples{
# Load the dataset & folds
MG.44

# Access the dataset
MG.44$Dataset

# Access the folds
MG.44$Folds

# Access the coordinates
MG.44$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_MG.44 <- MG.44$Dataset[MG.44$Folds != 1,]
testing_data_MG.44 <- MG.44$Dataset[MG.44$Folds == 1,]

}
\keyword{datasets}
