% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MWP.36}
\alias{MWP.36}
\title{MWP.36 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
MWP.36
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, RSS
 \item Sample Size: 36
 \item Number of Features: 5
 \item Coordinates: With coordinates (EPSG: 32633)
 \item Location: Mecklenburg-Western Pomerania, Germany
 \item Sampling Design: Simple random sampling along field transects
 \item Study Area Size: 18 ha
 \item Geological Setting: Pleistocene young morainic landscape of the Weichselian glaciation with predominantly glacial sand
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Alexander Steiger (alexander.steiger@uni-rostock.de), University of Rostock
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): 28.02.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (28.02.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured through titration after oxidization of the organic carbon (Walkley & Black 1934)
   \item Sampling Date: August 2022
   \item Sampling Depth: 0 – 15 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in CaCl\eqn{_2} suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: August 2022
   \item Sampling Depth: 0 – 15 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: August 2022
   \item Sampling Depth: 0 – 15 cm
   }

}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from “LAiV Geodaten-MV”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sampling Date: Unknown
  }

  RSS – Remote Sensing Derived Spectral Data
  \itemize{
  \item Number of Features: 3
  \item Code(s): \code{B02, B8A, B11}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 bare soil image (Level-2A) from “Copernicus Open Access Hub”, with bands of 10 - 20 m spatial resolution
  \item Processing: Extracting RSS values from raster at soil sampling locations, selecting bands spread throughout the spectral range with lower intercorrelation due to low sample size
  \item Sampling Date: August 2022
  }

}
\cr

\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe With coordinates (EPSG: 32633)
}
\examples{
# Load the dataset list
MWP.36

# Access the dataset
MWP.36$Dataset

# Access the folds
MWP.36$Folds

# Access the coordinates
MWP.36$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_MWP.36 <- MWP.36$Dataset[MWP.36$Folds != 1,]
testing_data_MWP.36 <- MWP.36$Dataset[MWP.36$Folds == 1,]

}
\references{
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.\cr
\cr
Walkley, A. & Black, I. A. (1934). An examination of the Degtjareff method for determining soil organic matter, and a proposed modification of the chromic acid titration method. Soil science, 37(1), 29-38.
}
\keyword{datasets}
