% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRW.115}
\alias{NRW.115}
\title{NRW.115 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
NRW.115
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: MIR
 \item Sample size: 115
 \item Number of Features: 1,686
 \item Coordinates: Without coordinates because of privacy concerns
 \item Location: North Rhine-Westphalia, Germany
 \item Sampling Design: Regular Grid Sampling on two adjacent fields in two sampling campaigns
 \item Study Area Size: 17 ha
 \item Geological Setting: Cretaceous marls partially covered by Saalian glacial till, aeolian sand and fluvial sediments
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Stefan Paetzold (s.paetzold@uni-bonn.de), University of Bonn
     }
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Determined by the difference of total carbon and inorganic carbon, where total carbon was obtained through elemental analysis by measuring the CO\eqn{_2} release during dry combustion (DIN ISO 10694) without acid pretreatment and inorganic carbon as 0.12 x the calcium carbonate content, determined by the gas-volumetric Scheibler Method (ISO 10693)
   \item Sampling Date: October 2013 & November 2015
   \item Sampling Depth: 0 - 30 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in CaCl\eqn{_2} suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: October 2013 & November 2015
   \item Sampling Depth: 0 - 30 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Measured through a combination of wet sieving and sedimentation after dispersion and removal of organic matter through oxidation (DIN ISO 11277)
   \item Sampling Date: October 2013 & November 2015
   \item Sampling Depth: 0 - 30 cm
   }
}
\cr
Groups of Features:
\describe{
  MIR – Mid Infrared Spectroscopy
  \itemize{
  \item Number Features: 1,686 \cr
  \item Code(s): \code{wn_3799, wn_3797.1, wn_3795.1} ... \code{wn_549.6} \cr
  \item Unit: Unitless \cr
  \item Sensing: MIR spectrometer (Bruker Tensor 27 HTS-XT, Bruker Optik, Ettlingen, Germany), on dried and sieved samples (<2 mm) in the laboratory, spectral range was 7,500 – 549.6 cm\eqn{^{-1}} at 4 cm\eqn{^{-1}} intervals \cr
  \item Processing: Discarding irrelevant spectral data of the spectrum (7,500 - 3,799 cm\eqn{^{-1}}), resampling to ~2 cm\eqn{^{-1}} intervals\cr
  \item Sensing Date: October 2013 & November 2015
  \item Spectral Information (after data processing):
   \itemize{
     \item Data Representation: Wavenumber (in cm\eqn{^{-1}})
     \item Measurement Type: Reflectance
     \item Spectral Resolution: ~ 2 cm\eqn{^{-1}}
     \item Spectral Range: 3,799 - 549.6 cm\eqn{^{-1}}
   }
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset list
NRW.115

# Access the dataset
NRW.115$Dataset

# Access the folds
NRW.115$Folds

# Access the coordinates but empty for NRW.115 (i.e. NA)
NRW.115$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_NRW.115 <- NRW.115$Dataset[NRW.115$Folds != 1,]
testing_data_NRW.115 <- NRW.115$Dataset[NRW.115$Folds == 1,]

}
\keyword{datasets}
