% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSW.52}
\alias{NSW.52}
\title{NSW.52 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
NSW.52
}
\description{
\itemize{
 \item Target Soil Properties: SOM_target, pH_target, Cay_target
 \item Groups of Features: RSS, DEM
 \item Sample size: 52
 \item Number of Features: 6
 \item Coordinates: With coordinates (EPSG: 32755)
 \item Location: New South Wales, Australia
 \item Sampling Design: K-means (30 samples) & XXX (22 sampples)
 \item Study Area: 1,070 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: patrick.filippi@sydney.edu.au
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOM_target}}{XXX}
  \item{\code{clay_target}}{XXX}
  \item{\code{pH_target}}{XXX} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{RSS}}{
  No. Features: 3 \cr
  Feature Code(s): B04, B08, B11 \cr
  Unit: unitless \cr
  Sensing: Sentinel-2 bare soil Image (Level-2A), with selected bands based on correlation\cr
  Processing: Extracting RSS values from raster at soil sampling locations \cr
  Sensing Date: July 2018
  }
  \item{\code{DEM}}{
  No. Features: 3 \cr
  Feature Code(s): Altitude, Slope, TDI \cr
  Unit: Altitude (m), Slope (°), TDI (unitless) \cr
  Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from the “Elevation and Depth – Foundation Spatial Data (ELVIS)”\cr
  Processing: Slope and TPI through \code{\link[raster]{terrain}} function of the \pkg{raster} package \cr
  Sensing Date: April 2016
  }
}


\strong{Folds:}
Vector with 52 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32755)
}
\examples{
# Load the dataset & folds
NSW.52

# Access the dataset
NSW.52$Dataset

# Access the folds
NSW.52$Folds

# Access the coordinates
NSW.52$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_NSW.52 <- NSW.52$Dataset[NSW.52$Folds != 1,]
testing_data_NSW.52 <- NSW.52$Dataset[NSW.52$Folds == 1,]

}
\keyword{datasets}
