% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{O.32}
\alias{O.32}
\title{O.32 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
O.32
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: MIR
 \item Sample Size: 32
 \item Number of Features: 1,637
 \item Coordinates: Without coordinates because of privacy concerns
 \item Location: Occitanie, France
 \item Sampling Design: Regular grid sampling
 \item Study Area Size: 1.5 ha
 \item Geological Setting: Pleistocene fluvial deposits and Miocene marine deposits
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Stefan Paetzold (s.paetzold@uni-bonn.de), University of Bonn
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Determined by the difference of total carbon and inorganic carbon, where total carbon was obtained through elemental analysis by measuring the CO\eqn{_2} release during dry combustion (DIN ISO 10694) without acid pretreatment and inorganic carbon as 0.12 x the calcium carbonate content, determined by the gas-volumetric Scheibler Method (ISO 10693)
   \item Sampling Date: March 2020
   \item Sampling Depth: 0 - 20 cm
   }


   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in CaCl\eqn{_2} suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: March 2020
   \item Sampling Depth: 0 - 20 cm
}


   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water, German adaptation (DIN ISO 11277)
   \item Sampling Date: March 2020
   \item Sampling Depth: 0 - 20 cm
   }

}

\cr
Groups of Features:
\describe{
  MIR – Mid Infrared Spectroscopy
  \itemize{
  \item Number of Features: 1,637
  \item Code(s): \code{wn_3800, wn_3797.1, wn_3795.1} ... \code{wn_751.1} \cr
  \item Unit: \% (Reflectance)
  \item Sensing: Handheld MIR spectrometer (Agilent 4300, Agilent Technologies, Santa Clara, USA), on dried and sieved samples (<2 mm) in the laboratory, spectral range was 5,200 – 650 cm\eqn{^{-1}} at 4 - 16 cm\eqn{^{-1}} intervals \cr
  \item Processing: Discarding irrelevant spectral data of the spectrum (5,200 - 3,800 cm\eqn{^{-1}}) and noisy edges of the spectrum (751.1 - 650 cm\eqn{^{-1}}), resampling to ~2 cm\eqn{^{-1}} intervals\cr
  \item Sampling Date: March 2020
  \item Spectral Information (After Data Processing):
   \itemize{
     \item Data Representation: Wavenumber (in cm\eqn{^{-1}})
     \item Spectral Resolution: ~2 cm\eqn{^{-1}}
     \item Spectral Range: 3,800 - 751.1 cm\eqn{^{-1}}
   }
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset list
O.32

# Access the dataset
O.32$Dataset

# Access the folds
O.32$Folds

# Access the coordinates but empty for O.32 (i.e. NA)
O.32$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_O.32 <- O.32$Dataset[O.32$Folds != 1,]
testing_data_O.32 <- O.32$Dataset[O.32$Folds == 1,]

}
\keyword{datasets}
