% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Overview_Datasets}
\alias{Overview_Datasets}
\title{Overview of Datasets in LimeSoDa}
\format{
An object of class \code{data.frame} with 31 rows and 8 columns.
}
\usage{
Overview_Datasets
}
\description{
This table gives an overview of datasets used in LimeSoDa and containts:
\itemize{
 \item Dataset_ID: ID (i.e. name) of the dataset
 \item Number_of_samples: The number of analysed soil samples (i.e. rows) in the dataset
 \item Number_of_features: The number of features (i.e. feature columns) in the dataset
 \item Sensors: The type of sensors used to create the features in the dataset, see abbrevations below
 \item Coordinates: Specifies whether coordinates or dummy covariates are available or excluded for the dataset
 \item Location: The location where the dataset was collected
 \item Study_area_in_ha: The size of the study area in ha
 \item Sampling_Design: The sampling design used to collect the dataset as short description compared to metadata description
 }
 Abbreviations for sensors:
\itemize{
 \item CSMoisture - Capacitive soil moisture
 \item DEM - Digital elevation model and terrain parameters
 \item ERa - Apparent electrical resistivity
 \item Gamma - Gamma-ray activity
 \item MIR - Mid infrared spectroscopy
 \item NIR - Near infrared spectroscopy
 \item pH-ISE - Ion selective electrodes for pH determination
 \item RSS - Remote sensing derived spectral data
 \item VI - Vegetation indices
 \item vis-NIR - Visible and near infrared spectroscopy
 \item XRF - X-ray fluorescence derived elemental concentrations
 }
}
\keyword{datasets}
