% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP.62}
\alias{RP.62}
\title{RP.62 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
RP.62
}
\description{
\itemize{
 \item Target Soil Properties: SOC_target, pH_target, Clay_target
 \item Groups of Features: ER, Gamma, ISE.pH, NIR
 \item Sample size: 62
 \item Number of Features: 1408
 \item Coordinates: With coordinates (EPSG: 32632)
 \item Location: Rhineland-Palatinate, Germany
 \item Sampling Design: Regular Grid Sampling
 \item Study Area: 3.3 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: SVogel@atb-potsdam.de
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOC_target}}{Unit: \% \cr
   Protocol: Dry combustion after removing inorganic carbon (DIN ISO 10694)\cr
   Sampling Date: XXX 2017\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{clay_target}}{Unit: \% \cr
   Protocol: Wet sieving and sedimentation after removal of organic matter (DIN ISO 11277)\cr
   Sampling Date: XXX 2017\cr
   Sampling Depth: 0 - 30 cm}
  \item{\code{pH_target}}{Unit: Unitless \cr
   Protocol: In CaCl\eqn{_2} solution with a glass electrode (DIN ISO 10390)\cr
   Sampling Date: September 2020\cr
   Sampling Depth: 0 - 30 cm} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{ER}}{
  No. Features: 1 \cr
  Feature Code(s): ER \cr
  Unit: \eqn{\Omega} m \cr
  Sensing: Array of multiple rolling electrodes (Geophilus company, Caputh, Germany) on RapidMapper platform with 0 - 50 cm measuring depth (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: September 2020
  }

  \item{\code{pH_ISE}}{
  No. Features: 1 \cr
  Feature Code(s): pH_ISE \cr
  Unit: unitless \cr
  Sensing: Soil pH Manager (VERIS Technologies, Salinas, USA ) on RapidMapper platform (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: September 2020
  }

  \item{\code{Gamma}}{
  No. Features: 5 \cr
  Feature Code(s): G_Total_Counts, G_K, G_U, G_Th, G_Cs \cr
  Unit: unitless \cr
  Sensing: Gamma sensor (MS-2000-CsI-MTS, Medusa Radiometrics BV, Groningen, Netherlands) on RapidMapper platform (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations \cr
  Sensing Date: September 2020
  }
  \item{\code{NIR}}{
  No. Features: 1401 \cr
  Feature Code(s): NIR_1000, NIR_1001, NIR_1002 ... NIR_2400 \cr
  Unit: unitless \cr
  Sensing: NIR spectrometer (C11118GA, Hamamatsu Photonics K.K., Shizuoka Prefecture, Japan) on RapidMapper platform (in-situ)\cr
  Processing: Kriging to align sensing- with soil sampling locations, Harmonization to 1nm intervals, Discarding noisy edges \cr
  Sensing Date: September 2020
  }
}


\strong{Folds:}
Vector with 62 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32632)
}
\examples{
# Load the dataset & folds
RP.62

# Access the dataset
RP.62$Dataset

# Access the folds
RP.62$Folds

# Access the coordinates
RP.62$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_RP.62 <- RP.62$Dataset[RP.62$Folds != 1,]
testing_data_RP.62 <- RP.62$Dataset[RP.62$Folds == 1,]

}
\keyword{datasets}
