% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC.50}
\alias{SC.50}
\title{SC.50 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
SC.50
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa
 \item Sample Size: 50
 \item Number of Features: 3
 \item Coordinates: With coordinates (EPSG: 32722)
 \item Location: Santa Catarina, Brazil
 \item Sampling Design: Regular grid sampling
 \item Study Area Size: 13 ha
 \item Geological Setting: Heavily weathered soils originating from Mesozoic basalt rocks
 \item Previous Data Publication: Full dataset published in Bottega & Safanelli (2024)
 \item Contact Information:
   \itemize{
     \item Eduardo Bottega (bottega.elb@gmail.com), Federal University of Santa Maria
     \item José Lucas Safanelli (jsafanelli@woodwellclimate.org), Woodwell Climate Research Center
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured through titration after oxidization of the organic carbon (Walkley & Black 1934)
   \item Sampling Date: November 2013
   \item Sampling Depth: 0 – 20 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio
   \item Sampling Date: November 2013
   \item Sampling Depth: 0 – 20 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water, German adaptation (DIN ISO 11277)
   \item Sampling Date: November 2013
   \item Sampling Depth: 0 – 20 cm
   }

}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (30 m) based on synthetic aperture radar from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sampling Date: October 2011
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: LandMapper ERM-02 conductivity meter (Landviser, League City, USA) with exploration depth of 0 - 20 cm, in-situ
  \item Processing: None
  \item Sampling Date: November 2014
  }

}
\cr

\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32722)
}
\examples{
# Load the dataset list
SC.50

# Access the dataset
SC.50$Dataset

# Access the folds
SC.50$Folds

# Access the coordinates
SC.50$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SC.50 <- SC.50$Dataset[SC.50$Folds != 1,]
testing_data_SC.50 <- SC.50$Dataset[SC.50$Folds == 1,]

}
\references{
Bottega, E. L. & Safanelli J. L. (2024). Data for "Site-Specific Management Zones Delineation Based on Apparent Soil Electrical Conductivity in Two Contrasting Fields of Southern Brazil". Zenodo repository.  https://doi.org/10.5281/zenodo.13770031 \cr
\cr
Walkley, A. & Black, I. A. (1934). An examination of the Degtjareff method for determining soil organic matter, and a proposed modification of the chromic acid titration method. Soil science, 37(1), 29-38.
}
\keyword{datasets}
