% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC.93}
\alias{SC.93}
\title{SC.93 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
SC.93
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: vis-NIR
 \item Sample size: 93
 \item Number of Features: 2,146
 \item Coordinates: With coordinates (EPSG: 32722)
 \item Location: Santa Catarina, Brazil
 \item Sampling Design: conditioned Latin Hypercube Sampling based on terrain parameters
 \item Study Area Size: 108 ha
 \item Geological Setting: Heavily weathered soils originating from volcanic rock of the Serra Geral Formation (basalt, dacit)
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Taciara Zborowski Horst (taciaraz@utfpr.edu.br), Federal University of Technology – Paraná
     \item Ricardo Simão Diniz Dalmolin (dalmolin@ufsm.br), Federal University of Santa Maria
     }
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured through light absorption after oxidization of the organic carbon in suspension (Tedesco et al. 1995)
   \item Sampling Date: December 2016
   \item Sampling Depth: 0 - 20 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode ratio with a 1:1 liquid:soil [? volmetric or gravimetric ?] ratio
   \item Sampling Depth: 0 - 20 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: December 2016
   \item Sampling Depth: 0 - 20 cm
   }
}

\cr
Groups of Features:
\describe{
  vis-NIR – Visible and Near Infrared Spectroscopy
  \itemize{
  \item Number Features: 2,146 \cr
  \item Code(s): \code{wl_355, wl_356, wl_357} ... \code{wl_2500} \cr
  \item Unit: Unitless \cr
  \item Sensing: vis-NIR spectrometer (ASD FieldSpec 4, Analytical Spectral Devices Inc., Boulder, USA), on dried and sieved samples (<2 mm) in the laboratory, spectral range was 355 - 2,500 nm at 3 - 8 nm spectral resolution
  \item Processing: Reesampling to 1 nm intervals\cr
  \item Sensing Date: March 2017
  \item Spectral Information (after data processing):
   \itemize{
     \item Data Representation: Wavelength (in nm)
     \item Measurement Type: Reflectance
     \item Spectral Resolution: 1 nm
     \item Spectral Range: 355 – 2500 nm
   }
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32722)
}
\examples{
# Load the dataset list
SC.93

# Access the dataset
SC.93$Dataset

# Access the folds
SC.93$Folds

# Access the coordinates
SC.93$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SC.93 <- SC.93$Dataset[SC.93$Folds != 1,]
testing_data_SC.93 <- SC.93$Dataset[SC.93$Folds == 1,]

}
\references{
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.\cr
\cr
Tedesco, M.J., Gianello, C., Bissani, C., Bohnen, H. & Volkweiss, S.J. (1995) Análise de solo, plantas e outros materiais. [Analysis of soil, plants and other materials.] 2nd Edition, Departamento de Solos da Universidade Federal do Rio Grande do Sul, Porto Alegre, 174.
}
\keyword{datasets}
