% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SM.40}
\alias{SM.40}
\title{SM.40 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
SM.40
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa
 \item Sample size: 40
 \item Number of Features: 3
 \item Coordinates: With coordinates (EPSG: 32633)
 \item Location: South Moravia, Czechia
 \item Sampling Design: Subsample from previous Regular Grid Sampling, subsample was handpicked to cover contrasting areas
 \item Study Area Size: 53 ha
 \item Geological Setting: Weichselian sandy loess [? Eventually wrong ?]
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Vojtech Lukas (vojtech.lukas@mendelu.cz), Mendel University in Brno
     }
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOM_target}
   \item Unit: \%
   \item Protocol: Measured through titration after oxidization of the organic carbon following slight adjustments of the Walkley & Black (1934) method (Zbíral et al. 2004)
   \item Sampling Date: May 2004
   \item Sampling Depth: 0 - 30 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in KCl suspension with a glass electrode with unspecified [?please add if you the volumetric or gravimetric soil to liquid ratio?] liquid:soil ratio (Zbíral 2002)
   \item Sampling Date: May 2004
   \item Sampling Depth: 0 - 30 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Hydrometer method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by measuring suspension density using a hydrometer following slight adjustments of the Bouyoucos (1927) method [? Citation ?]
   \item Sampling Date: April 2006
   \item Sampling Depth: 0 - 30 cm
   }
}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} (m), \code{Slope } (°)
  \item Sensing: Digital elevation model raster (~2 m) based on LiDAR from “Geoportal of the Czech Office for Surveying, Mapping and Cadastre"
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sensing Date: Unknown
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: EM38 sensor (Geonics Ltd., Mississauga, Canada) drawn by a vehicle with exploration depth of 0 - 75 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sensing Date: May 2004
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe With coordinates (EPSG: 32633)
}
\examples{
# Load the dataset list
SM.40

# Access the dataset
SM.40$Dataset

# Access the folds
SM.40$Folds

# Access the coordinates
SM.40$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SM.40 <- SM.40$Dataset[SM.40$Folds != 1,]
testing_data_SM.40 <- SM.40$Dataset[SM.40$Folds == 1,]

}
\references{
Bouyoucos, G. J. (1927). The hydrometer as a new method for the mechanical analysis of soils. Soil science, 23(5), 343-354.\cr
\cr
Walkley, A. & Black, I. A. (1934). An examination of the Degtjareff method for determining soil organic matter, and a proposed modification of the chromic acid titration method. Soil science, 37(1), 29-38. \cr
\cr
Zbíral, J., Honsa, I., Malý, S. & Čižmář, D (2004). Analýza půd III : jednotné pracovní postupy [Soil Analysis III : Unified working procedures]. Brno: UKZUZ, 199\cr
\cr
Zbíral, J. (2002). Analýza půd I : jednotné pracovní postupy [Soil analysis I: Integrated work procedures]. Brno: UKZUZ, 197
}
\keyword{datasets}
