% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSP.58}
\alias{SSP.58}
\title{SSP.58 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
SSP.58
}
\description{
\itemize{
 \item Target Soil Properties: SOM_target, pH_target, Cay_target
 \item Groups of Features: visNIR
 \item Sample size: 58
 \item Number of Features: 351
 \item Coordinates: Without coordinates because dataset was not georeferenced
 \item Location: Sao Paulo State, Brazil
 \item Sampling Design: Regular Grid Sampling
 \item Study Area: 0.7 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: tiagosrt@usp.br
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOM_target}}{XXX}
  \item{\code{clay_target}}{XXX}
  \item{\code{pH_target}}{XXX} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{visNIR}}{XXX
  }
}


\strong{Folds:}
Vector with 58 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset & folds
SSP.58

# Access the dataset
SSP.58$Dataset

# Access the folds
SSP.58$Folds

# Access the coordinates but empty for SSP.58 (i.e. NA)
SSP.58$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SSP.58 <- SSP.58$Dataset[SSP.58$Folds != 1,]
testing_data_SSP.58 <- SSP.58$Dataset[SSP.58$Folds == 1,]

}
\keyword{datasets}
