% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UL.120}
\alias{UL.120}
\title{UL.120 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
UL.120
}
\description{
\itemize{
 \item Target Soil Properties: SOM, pH, Clay
 \item Groups of Features: ERa, vis-NIR
 \item Sample Size: 120
 \item Number of Features: 2,082
 \item Coordinates: Without coordinates because of privacy concerns instead with dummy coordinates (EPSG: 4326)
 \item Location: Uppsala Län, Sweden
 \item Sampling Design: Three sampling designs over multiple adjacent fields: (1) regular grid sampling, targeted sampling through surface tortoise sampling (Persson et al. 2023) based on (2) ERa and (3) reflectance from remote sensing
 \item Study Area Size: 97 ha
 \item Geological Setting: Glacial and postglacial clay with elements of sandy till
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Johanna Wetterlind (Johanna.Wetterlind@slu.se), Swedish University of Agricultural Sciences
     \item Bo Stenberg (Bo.Stenberg@slu.se), Swedish University of Agricultural Sciences
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): 28.02.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (28.02.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOM - Soil Organic Matter
   \itemize{
   \item Code: \code{SOM_target}
   \item Unit: \%
   \item Protocol: Measured through the weight difference before and after ignition of the soil with additional correction for structural water from clay by using the formula: SOM = LI − 0.46 − 0.047 × clay content (\%) (Ekstroem 1927)
   \item Sampling Date: September 2005
   \item Sampling Depth: 0 - 20 cm
   }


   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in a water suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio
   \item Sampling Date: September 2005
   \item Sampling Depth: 0 - 20 cm
}


   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: September 2005
   \item Sampling Depth: 0 - 20 cm
   }
}
\cr
Groups of Features:
\describe{
  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: EM38 sensor (Geonics Ltd., Mississauga, Canada) with exploration depth of 0 - 150 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: April 2005
  }


  vis-NIR – Visible and Near Infrared Spectroscopy
  \itemize{
  \item Number of Features: 2,081
  \item Code(s): \code{wl_420, wl_421, wl_422} ... \code{wl_2500}
  \item Unit: \% (Reflectance)
  \item Sensing: FieldSpec Pro FR scanning instrument (Analytical Spectral Devices Inc., Boulder, USA), on dried and sieved samples (<2 mm) in the laboratory, spectral range was 350 – 2,500 nm at 1.4 – 2.0 nm intervals
  \item Processing: Discarding noisy edges of the spectrum (350 - 420 nm), resampling to 1 nm intervals
  \item Sampling Date: September 2005
  \item Spectral Information (After Data Processing):
   \itemize{
     \item Data Representation: Wavelength (in nm)
     \item Spectral Resolution: 1 nm
     \item Spectral Range: 420 - 2,500 nm
   }
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with dummy coordinates (EPSG: 4326)
}
\examples{
# Load the dataset list
UL.120

# Access the dataset
UL.120$Dataset

# Access the folds
UL.120$Folds

# Access the coordinates but for UL.120 only dummy coordinates
UL.120$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_UL.120 <- UL.120$Dataset[UL.120$Folds != 1,]
testing_data_UL.120 <- UL.120$Dataset[UL.120$Folds == 1,]

}
\references{
Ekstroem, G. (1927). Klassifikation av Svenska Åkerjordar (Classification of Swedish arable soils). Sveriges Geologiska Undersoekning, Ser C. 345, 161 pp. \cr
\cr
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.\cr
\cr
Persson, K., Soederstroem, M. & Mutua, J. (2023). SurfaceTortoise: Find Optimal Sampling Locations Based on Spatial Covariate(s). R package version 2.0.1.
}
\keyword{datasets}
