% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W.50}
\alias{W.50}
\title{W.50 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
W.50
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa, VI, XRF
 \item Sample Size: 50
 \item Number of Features: 15
 \item Coordinates: Without coordinates because of privacy concerns
 \item Location: Wisconsin, USA
 \item Sampling Design: Conditioned latin hypercube sampling based on electrical conductivity, terrain parameters, and normalized difference vegetation index
 \item Study Area Size: 80 ha
 \item Geological Setting: Glacial outwash and sediments of the Johnson End Moraine
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Jingyi Huang (jhuang426@wisc.edu), University of Wisconsin-Madison
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured CO\eqn{_2} release during dry combustion after removing inorganic carbon with an acid (Nelson and Sommers 1996)
   \item Sampling Date: July 2019
   \item Sampling Depth: 0 – 10 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode with a 1:1 liquid:soil gravimetric ratio (Burt 2014)
   \item Sampling Date: July 2019
   \item Sampling Depth: 0 – 10 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Hydrometer method; separation of the fractions by sieving and sedimentation. Measurement of the separated fractions by weighing the density of the suspension (Gee and Bauder 1979)
   \item Sampling Date: July 2019
   \item Sampling Depth: 0 – 10 cm
   }

}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (3 m) based on LiDAR from the “Wisconsin Department of Natural Resources”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations, resampled from the original 3 m resolution to 5 m resolution
  \item Sampling Date: Unknown
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: DUALEM-1HS instrument (DUALEM Inc., Milton, Canada) with exploration depth of 0 - 30 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: July 2019
  }

  VI - Vegetation Indices
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{NDVI, GNDVI}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 Image during vegetative period (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{NDVI} as (B08 - B04) / (B08 + B04) and \code{GNDVI} as (B08 - B03) / (B08 + B03), extracting VI values from raster at soil sampling locations
  \item Sampling Date: July 2019
   }

  XRF – X-ray Fluorescence Derived Elemental Concentrations
  \itemize{
  \item Number of Features: 10
  \item Code(s): \code{XRF_Mg, XRF_Al, XRF_Si,  XRF_Ca, XRF_Ti, XRF_Mn, XRF_Fe, XRF_Zn, XRF_Sr, XRF_Zr}
  \item Unit: ppm (estimated through XRF Geochem not ground truth)
  \item Sensing: Delta Premium PXRF spectrometer (Olympus Scientific Solutions Americas Inc., Waltham, USA), on dried and sieved samples (<2 mm) in the laboratory
  \item Processing: Compton normalization method to transform full spectra into estimates of elemental concentrations with accompanied software of the sensor (Geochem mode)
  \item Sampling Date: July 2019
  }

}
\cr

\strong{Folds:}
 Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset list
W.50

# Access the dataset
W.50$Dataset

# Access the folds
W.50$Folds

# Access the coordinates but empty for W.50 (i.e. NA)
W.50$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_W.50 <- W.50$Dataset[W.50$Folds != 1,]
testing_data_W.50 <- W.50$Dataset[W.50$Folds == 1,]

}
\references{
Burt, R. (Ed.) (2014). Kellogg soil survey laboratory methods manual. United States Department of Agriculture, Natural Resources Conservation Service, National Soil Survey Center, Kellogg Soil Survey Laboratory.\cr
\cr
Gee, G. W., & Bauder, J. W. (1979). Particle size analysis by hydrometer: a simplified method for routine textural analysis and a sensitivity test of measurement parameters. Soil Science Society of America Journal, 43(5), 1004-1007.\cr
\cr
Nelson, D.W. & Sommers, L.E. (1996) Total Carbon, Organic Carbon, and Organic Matter. In: Sparks, D.L., Page, A.L., Helmke, P.A., Loeppert, R.H., Soltanpour, P.N., Tabatabai, M.A., Johnston, C.T. & Sumner, M.E., Eds., Methods of Soil Analysis. Part 3. Chemical Methods, Soil Science Society of America, Madison, WI, 961-1010.
}
\keyword{datasets}
