% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BB.51}
\alias{BB.51}
\title{BB.51 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
BB.51
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa, pH-ISE
 \item Sample size: 51
 \item Number of Features: 4
 \item Coordinates: With coordinates (EPSG: 25833)
 \item Location: Brandenburg, Germany
 \item Sampling Design: Multi Criteria Sampling based on quantile coverage of the sensing-features (ECa and pH_ISE), clustering of large and low values of sensing-features and spatial coverage
 \item Study Area Size: 40 ha
 \item Geological Setting: Pleistocene young morainic landscape of the Weichselian glaciation with predominantly glacial sand
 \item Previous Data Publication: Target soil properties published but under boycott in Vogel et al. 2022
 \item Contact Information:
   \itemize{
     \item Sebastian Vogel (SVogel@atb-potsdam.de), Leibniz Institute for Agricultural Engineering and Bioeconomy (ATB)
     }
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Measured CO\eqn{_2} release during dry combustion after removing inorganic carbon with an acid (DIN ISO 10694)
   \item Sampling Date: September - November 2017
   \item Sampling Depth: 0 - 30 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in CaCl\eqn{_2} suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: September - November 2017
   \item Sampling Depth: 0 - 30 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Measured through a combination of wet sieving and sedimentation after dispersion and removal of organic matter through oxidation (DIN ISO 11277)
   \item Sampling Date: September - November 2017
   \item Sampling Depth: 0 - 30 cm
   }
}

\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} (m), \code{Slope } (°)
  \item Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from “GeoBasis-DE/LGB”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sensing Date: LiDAR March 2009, images for photogrammetry May 2018
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: EC Surveyor (VERIS Technologies, Salinas, USA) from VerisMSP3 with exploration depth of 0 - 30 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sensing Date: September 2017
  }

  pH_ISE – Ion Selective Electrodes for pH Determination
  \itemize{
  \item Number Features: 1
  \item Code(s): \code{pH_ISE}
  \item Unit: Unitless
  \item Sensing: Soil pH Manager (VERIS Technologies, Salinas, USA) from VerisMSP3, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sensing Date: September 2017
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 25833)
}
\examples{
# Load the dataset list
BB.51

# Access the dataset
BB.51$Dataset

# Access the folds
BB.51$Folds

# Access the coordinates
BB.51$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_BB.51 <- BB.51$Dataset[BB.51$Folds != 1,]
testing_data_BB.51 <- BB.51$Dataset[BB.51$Folds == 1,]

}
\references{
Vogel S., Bönecke E., Kling C., Kramer E., Lück K., Nagel A., Philipp G., Rühlmann J., Schröter I. & Gebbers, R. (2022), Base neutralizing capacity from agricultural fields in the quaternary landscape of North-East Germany, BONARES Repository, https://doi.org/10.20387/bonares-zh3x-nd80
}
\keyword{datasets}
