% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H.138}
\alias{H.138}
\title{H.138 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
H.138
}
\description{
\itemize{
METADATA NOT YET RECEIVED ONLY PRILIMINARY ENTRIES
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: MIR
 \item Sample size: 138
 \item Number of Features: 2,489
 \item Coordinates: XXX
 \item Location: Hubei, China
 \item Sampling Design: XXX
 \item Study Area Size: 420 ha
 \item Geological Setting: Sedimentary rocks, mainly dolomite with silt and limestone formed in the middle and lower Jurassic
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Alexandre M.J.-C- Wadoux  (Alexandre.Wadoux@inrae.fr), University of Montpellier
     }
 \item License: CC BY-NC-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: XXX
   \item Sampling Date: June 2013, May, 2014 & November 2014
   \item Sampling Depth: 0 - 20 cm
   }


   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: XXX
   \item Sampling Date: June 2013, May, 2014 & November 2014
   \item Sampling Depth: 0 - 20 cm
}


   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: XXX
   \item Sampling Date: June 2013, May, 2014 & November 2014
   \item Sampling Depth: 0 - 20 cm
   }
}
\cr
Groups of Features:
\describe{
  MIR – Mid Infrared Spectroscopy
  \itemize{
  \item Number Features: 2,489 \cr
  \item Code(s): \code{wn_5397.9, wn_5396, wn_5394} ... \code{wn_599.8}
  \item Unit: Unitless
  \item Sensing: XXX
  \item Processing: XXX
  \item Sensing Date: June 2013, May, 2014 & November 2014
  \item Spectral Information (after data processing):
   \itemize{
     \item Data Representation: Wavenumber (in cm\eqn{^{-1}})
     \item Measurement Type: Reflectance
     \item Spectral Resolution: ~2 cm\eqn{^{-1}}
     \item Spectral Range: 5,397.9 – 599.8 cm\eqn{^{-1}}
   }
  }

}
\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
XXX
}
\examples{
# Load the dataset list
H.138

# Access the dataset
H.138$Dataset

# Access the folds
H.138$Folds

# Access the coordinates
H.138$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_H.138 <- H.138$Dataset[H.138$Folds != 1,]
testing_data_H.138 <- H.138$Dataset[H.138$Folds == 1,]

}
\references{
XXX
}
\keyword{datasets}
