% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SA.112}
\alias{SA.112}
\title{SA.112 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
SA.112
}
\description{
\itemize{
 \item Target Soil Properties: SOC, pH, Clay
 \item Groups of Features: DEM, ERa, Gamma, NIR, pH-ISE, VI
 \item Sample Size: 112
 \item Number of Features: 1,412
 \item Coordinates: Without coordinates because of privacy concerns
 \item Location: Saxony-Anhalt, Germany
 \item Sampling Design: Regular grid sampling but with missing values in the center of the field
 \item Study Area Size: 27 ha
 \item Previous Data Publication: None
 \item Geological Setting: Weichselian loess
 \item Contact Information:
   \itemize{
     \item Stefan Paetzold (s.paetzold@uni-bonn.de), University of Bonn
     \item Hamed Tavakoli (HTavakoli@atb-potsdam.de), Leibniz Institute for Agricultural Engineering and Bioeconomy (ATB)
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): XXX.03.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (XXX.03.2025): Initial release
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOC - Soil Organic Carbon
   \itemize{
   \item Code: \code{SOC_target}
   \item Unit: \%
   \item Protocol: Determined by the difference of total carbon and inorganic carbon, where total carbon was obtained through elemental analysis by measuring the CO\eqn{_2} release during dry combustion (DIN ISO 10694) without acid pretreatment and inorganic carbon as 0.12 x the calcium carbonate content, determined by the gas-volumetric Scheibler Method (ISO 10693)
   \item Sampling Date: October 2016
   \item Sampling Depth: 0 - 30 cm
   }

   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in a water suspension with a glass electrode with a 5:1 liquid:soil volumetric ratio (DIN ISO 10390)
   \item Sampling Date: October 2016
   \item Sampling Depth: 0 - 30 cm
}

   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water, German adaptation (DIN ISO 11277)
   \item Sampling Date: October 2016
   \item Sampling Depth: 0 - 30 cm
   }
}
\cr
Groups of Features:
\describe{
  DEM – Digital Elevation Model and Terrain Parameters
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{Altitude, Slope}
  \item Unit: \code{Altitude} in m, \code{Slope } in °
  \item Sensing: Digital elevation model raster (5 m) based on LiDAR and photogrammetry from “GeoBasis-DE / LVermGeo ST”
  \item Processing: Calculating \code{Slope} with \code{\link[raster]{terrain}} function of the \pkg{raster} package, extracting DEM values from raster at soil sampling locations
  \item Sampling Date: Unknown
  }

  ERa – Apparent Electrical Resistivity
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{ERa}
  \item Unit: \eqn{\Omega} m
  \item Sensing: Array of multiple rolling electrodes (Geophilus company, Caputh, Germany) on RapidMapper platform with exploration depth of 0 - 50 cm, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: August 2021
  }

  Gamma
  \itemize{
  \item Number of Features: 5
  \item Code(s): \code{G_Total_Counts, G_K, G_U, G_Th, G_Cs}
  \item Unit: Unitless
  \item Sensing: Passive gamma sensor (MS-2000-CsI-MTS, Medusa Radiometrics BV, Groningen, Netherlands) on RapidMapper platform, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: August 2021
  }

  NIR – Near Infrared Spectroscopy
  \itemize{
  \item Number of Features: 1,401
  \item Code(s): \code{wl_1000, wl_1001, wl_1002} ... \code{wl_2400}
  \item Unit: \% (Reflectance)
  \item Sensing: NIR spectrometer (C11118GA, Hamamatsu Photonics K.K., Shizuoka Prefecture, Japan) on RapidMapper platform, in-situ, spectral range was 900 - 2550 nm at 15 nm intervals
  \item Processing: Kriging to align sensing- with soil sampling locations, discarding noisy edges of the spectrum (900 - 1,000 nm & 2,400 - 2,550 nm), resampling to 1 nm intervals
  \item Sampling Date: August 2021
  \item Spectral Information (After Data Processing):
   \itemize{
     \item Data Representation: Wavelength (in nm)
     \item Spectral Resolution: 1 nm
     \item Spectral Range: 1,000 - 2,400 nm
   }
  }

  pH-ISE – Ion Selective Electrodes for pH Determination
  \itemize{
  \item Number of Features: 1
  \item Code(s): \code{pH-ISE}
  \item Unit: Unitless
  \item Sensing: Soil pH Manager (VERIS Technologies, Salinas, USA) on RapidMapper platform, in-situ
  \item Processing: Ordinary Kriging to align sensing- with soil sampling locations
  \item Sampling Date: August 2021
  }

  VI - Vegetation Indices
  \itemize{
  \item Number of Features: 2
  \item Code(s): \code{NDVI, GNDVI}
  \item Unit: Unitless
  \item Sensing: Sentinel-2 Image during vegetative period (Level-2A) from “Copernicus Open Access Hub”
  \item Processing: Calculating \code{NDVI} as (B08 - B04) / (B08 + B04) and \code{GNDVI} as (B08 - B03) / (B08 + B03), extracting VI values from raster at soil sampling locations
  \item Sampling Date: April 2018
  }
}
\cr
\strong{Folds:}
 Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Empty entry (NA)
}
\examples{
# Load the dataset list
SA.112

# Access the dataset
SA.112$Dataset

# Access the folds
SA.112$Folds

# Access the coordinates but empty for SA.112 (i.e. NA)
SA.112$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SA.112 <- SA.112$Dataset[SA.112$Folds != 1,]
testing_data_SA.112 <- SA.112$Dataset[SA.112$Folds == 1,]

}
\keyword{datasets}
