% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP.231}
\alias{SP.231}
\title{SP.231 Dataset}
\format{
A list with three elements: 'Dataset', 'Coordinates' and 'Folds'
}
\usage{
SP.231
}
\description{
\itemize{
 \item Target Soil Properties: SOM, pH, Clay
 \item Groups of Features: vis-NIR
 \item Sample Size: 231
 \item Number of Features: 271
 \item Coordinates: With coordinates (EPSG: 32654)
 \item Location: Saitama Prefecture, Japan
 \item Sampling Design: Two sampling designs over multiple fields depending on the soil conditions: (1) systematic sampling, in which samples are taken in the corners and middle of the field and (2) simple random sampling
 \item Study Area Size: 3.1 ha
 \item Geological Setting: Volcanic ash (Andosols)
 \item Previous Data Publication: None
 \item Contact Information:
   \itemize{
     \item Masakazu Kodaira (kodaira@cc.tuat.ac.jp), Tokyo University of Agriculture and Technology
     }
 \item License: CC BY-SA 4.0
 \item Publication/Modification Date (d/m/y): 28.02.2025, version 1.0
 \item Changelog:
   \itemize{
     \item Version 1.0 (28.02.2025): Initial release
     \item Version 1.1 (15.05.2025): Documentation for dimensionality bug fix
   }
}
\cr
}
\details{
\describe{
\strong{Dataset:}
Dataframe which contains the tabular dataset with the target soil properties and features}
\describe{
Target Soil Properties:

   SOM - Soil Organic Matter
   \itemize{
   \item Code: \code{SOM_target}
   \item Unit: \%
   \item Protocol: Measured through the weight difference before and after ignition
   \item Sampling Date: February 2017, December 2017 and February 2018
   \item Sampling Depth: 0 - 15 cm
   }


   pH
   \itemize{
   \item Code: \code{pH_target}
   \item Unit: Unitless
   \item Protocol: Measured in water suspension with a glass electrode with a 2.5:1 liquid:soil gravimetric ratio
   \item Sampling Date: February 2017, December 2017 and February 2018
   \item Sampling Depth: 0 - 15 cm
}


   Clay
   \itemize{
   \item Code: \code{Clay_target}
   \item Unit: \%
   \item Protocol: Sieve-Pipette method, measured through fractioning the soil into the sand fractions by sieving, and the silt and clay fractions by sedimentation in water (Gee and Bauder 1986)
   \item Sampling Date: February 2017, December 2017 and February 2018
   \item Sampling Depth: 0 - 15 cm
   }
}
\cr
Groups of Features:
\describe{
  vis-NIR – Visible and Near Infrared Spectroscopy
  \itemize{
  \item Number of Features: 271
  \item Code(s): \code{wl_350, wl_355, wl_360} ... \code{wl_1700}
  \item Unit: \% (Reflectance)
  \item Sensing: Mounted vis-NIR spectrometer (SAS3000, Shibuya Seiki Co. Ltd., Ehime Prefecture, Japan), in-situ, spectral range was 320 – 1,700 nm at 1 - 7 nm intervals
  \item Processing: Discarding noisy edges of the spectrum (320 - 350 nm), resampling to 5 nm intervals
  \item Sampling Date: February 2017, December 2017 and February 2018
  \item Spectral Information (After Data Processing):
   \itemize{
     \item Data Representation: Wavelength (in nm)
     \item Spectral Resolution: 1 nm
     \item Spectral Range: 350 - 1,700 nm
   }
  }

}

\cr


\strong{Folds:}
Vector which contains numerical entries from 1 to 10, each number can be used to index the folds for cross validation \cr

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32654)
}
\examples{
# Load the dataset list
SP.231

# Access the dataset
SP.231$Dataset

# Access the folds
SP.231$Folds

# Access the coordinates
SP.231$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SP.231 <- SP.231$Dataset[SP.231$Folds != 1,]
testing_data_SP.231 <- SP.231$Dataset[SP.231$Folds == 1,]

}
\references{
Gee, G.W. & Bauder, J.W. (1986) Particle-Size Analysis. In: Klute, A., Ed., Methods of Soil Analysis, Part 1. Physical and Mineralogical Methods, Agronomy Monograph No. 9, 2nd Edition, American Society of Agronomy/Soil Science Society of America, Madison, WI, 383-411.
}
\keyword{datasets}
