% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP.231}
\alias{SP.231}
\title{SP.231 Dataset}
\format{
A list with three elements: 'Dataset','Coordinates' and 'Folds'.
}
\usage{
SP.231
}
\description{
\itemize{
 \item Target Soil Properties: SOM_target, pH_target, Cay_target
 \item Groups of Features: visNIR, DEM
 \item Sample size: 231
 \item Number of Features: 272
 \item Coordinates: With coordinates (EPSG: 32654)
 \item Location: Saitama Prefecture, Japan
 \item Sampling Design: Random
 \item Study Area: 3.1 ha
 \item Publication/Modification Date (d/m/y): 02.08.2024, version 1.0
 \item Contact Information: kodaira@cc.tuat.ac.jp
 \item Extra Note: Pre-processing (e.g. Savitzky-Golay filter) needed for proper modelling, dimensionality reduction encouraged
}
\cr
}
\details{
\strong{Target Soil Properties:}
\describe{
  \item{\code{SOM_target}}{XXX}
  \item{\code{clay_target}}{XXX}
  \item{\code{pH_target}}{XXX} \cr
}

\strong{Groups of Features:}
\describe{
  \item{\code{visNIR}}{
  No. Features: 271 \cr
  Feature Code(s): XXX \cr
  Unit: unitless \cr
  Sensing: XXX\cr
  Processing: XXX \cr
  Sensing Date: XXX
  }
  \item{\code{DEM}}{
  No. Features: 1 \cr
  Feature Code(s): Altitude \cr
  Unit: Altitude (m) \cr
  Sensing: XXX \cr
  Processing: XXX \cr
  Sensing Date: XXX
  }
}


\strong{Folds:}
Vector with 52 entries in which numbers from 1 to 10 were randomly assigned to every entry. Each number refers to a fold out of the ten folds of equal size.

\strong{Coordinates:}
Dataframe with coordinates (EPSG: 32654)
}
\examples{
# Load the dataset & folds
SP.231

# Access the dataset
SP.231$Dataset

# Access the folds
SP.231$Folds

# Access the coordinates
SP.231$Coordinates

# How to split the dataset into training and testing folds for the example of the first fold
training_data_SP.231 <- SP.231$Dataset[SP.231$Folds != 1,]
testing_data_SP.231 <- SP.231$Dataset[SP.231$Folds == 1,]

}
\keyword{datasets}
